chore(refactor code)
- change css to scss
- Fix pagination bug
- fix email templates
- fix tests strings
- fix carousel on landing page
- fix toastr color
- change file name conventions

+++ server/__test__/recipes.spec.js
@@ -3,7 +3,7 @@ import supertest from 'supertest';
 import app from '../app';
 import recipes from './seed/recipeSeed';
 import reviews from './seed/reviewSeed';
-import users from './seed/userSeed';
+import users, { reviewUser } from './seed/userSeed';
 
 const request = supertest(app);
 const recipesApi = '/api/v1/recipes';
@@ -11,8 +11,8 @@ const recipesApi = '/api/v1/recipes';
 describe('Recipes Endpoint Tests', () => {
   describe('Test Recipe Creation', () => {
     it(
-      'Should return an error message when a user' +
-      'tries to create a recipe without being logged in',
+      `Should return an error message when a user
+ tries to create a recipe without being logged in`,
       (done) => {
         request.post(recipesApi)
           .set('Connection', 'keep alive')
@@ -48,8 +48,7 @@ describe('Recipes Endpoint Tests', () => {
         });
     });
     it(
-      'Should return an error message if no name' +
-      'is provided for recipe.',
+      'Should return an error message if no name is provided for recipe.',
       (done) => {
         const testUser = { ...users[0] };
         const testRecipe = { ...recipes[0] };
@@ -67,8 +66,8 @@ describe('Recipes Endpoint Tests', () => {
       }
     );
     it(
-      'Should return an error message if no description' +
-      'is provided for recipe.',
+      `Should return an error message
+ if no description is provided for recipe.`,
       (done) => {
         const testUser = { ...users[0] };
         const testRecipe = { ...recipes[0] };
@@ -86,8 +85,7 @@ describe('Recipes Endpoint Tests', () => {
       }
     );
     it(
-      'Should return an error message if no category is' +
-      'provided for recipe.',
+      'Should return an error message if no category is provided for recipe.',
       (done) => {
         const testUser = { ...users[0] };
         const testRecipe = { ...recipes[0] };
@@ -105,8 +103,7 @@ describe('Recipes Endpoint Tests', () => {
       }
     );
     it(
-      'Should return an error message if no ingredient' +
-      'is provided for recipe.',
+      'Should return an error message if no ingredient is provided for recipe.',
       (done) => {
         const testUser = { ...users[0] };
         const testRecipe = { ...recipes[0] };
@@ -126,8 +123,8 @@ describe('Recipes Endpoint Tests', () => {
       }
     );
     it(
-      'Should return an error message if no instruction' +
-      'is provided for recipe.',
+      `Should return an error message if
+ no instruction is provided for recipe.`,
       (done) => {
         const testUser = { ...users[0] };
         const testRecipe = { ...recipes[0] };
@@ -148,8 +145,8 @@ describe('Recipes Endpoint Tests', () => {
 
     describe('Test Modify recipe', () => {
       it(
-        'Should return an error if a user tries to' +
-        'edit a recipe without being logged in',
+        `Should return an error if a user tries
+ to edit a recipe without being logged in`,
         (done) => {
           const testRecipe = { ...recipes[0] };
           const recipeId = recipes[0].id;
@@ -168,8 +165,8 @@ describe('Recipes Endpoint Tests', () => {
         }
       );
       it(
-        'Should return an error message if the' +
-        'recipe to be edited is not found',
+        `Should return an error message
+ if the recipe to be edited is not found`,
         (done) => {
           const testUser = { ...users[0] };
           const testRecipe = { ...recipes[0] };
@@ -222,8 +219,8 @@ describe('Recipes Endpoint Tests', () => {
           }
         );
         it(
-          'Should return an error message if the recipe' +
-          'to be deleted does not exist',
+          `Should return an error message
+ if the recipe to be deleted does not exist`,
           (done) => {
             const testUser = { ...users[0] };
             request.delete(`${recipesApi}/${recipes[1]
@@ -239,8 +236,8 @@ describe('Recipes Endpoint Tests', () => {
         );
         describe('Review Endpoint Test', () => {
           it(
-            'Should prevent a non-logged in user' +
-            'from posting a review on a recipe',
+            `Should prevent a non-logged in
+ user from posting a review on a recipe`,
             (done) => {
               request.post(`${recipesApi}/${recipes[0].id}/reviews`)
                 .set('Connection', 'keep alive')
@@ -275,8 +272,8 @@ describe('Recipes Endpoint Tests', () => {
               });
           });
           it(
-            'Should return an error when a user tries' +
-            'to access a recipe that does not exist',
+            `Should return an error when a user
+ tries to access a recipe that does not exist`,
             (done) => {
               request
                 .post(`${recipesApi}/100/reviews?token=${users[0]
@@ -295,7 +292,7 @@ describe('Recipes Endpoint Tests', () => {
           it('Should add a review by a logged-in user', (done) => {
             request
               .post(`${recipesApi}/${recipes[0]
-                .id}/reviews?token=${users[0].tokens}`)
+                .id}/reviews?token=${reviewUser.tokens}`)
               .set('Connection', 'keep alive')
               .set('Content-Type', 'application/json')
               .type('form')
@@ -305,8 +302,8 @@ describe('Recipes Endpoint Tests', () => {
                 expect(res.body.reviewData).to.be.an('object');
                 expect(res.body.reviewData.content).to.be.a('string');
                 expect(res.body.message).to.equal('Review successfully posted');
-                done();
               });
+            done();
           });
           describe('Get Recipe Endpoint', () => {
             it('Should fetch a single recipe from the application', (done) => {
@@ -331,23 +328,27 @@ describe('Recipes Endpoint Tests', () => {
                   done();
                 });
             });
-            it('Should fetch all the user\'s recipe in the application', (done) => {
-              const testUser = { ...users[0] };
-              request.get(`/api/v1/users/recipes?token=${testUser.tokens}`)
-                .set('Connection', 'keep alive')
-                .set('Content-Type', 'application/json')
-                .type('form')
-                .end((err, res) => {
-                  expect(res.status).to.equal(200);
-                  expect(res.body.message)
-                    .to.equal(`You currently have ${res
-                      .body.recipeData.length} recipe(s)`);
-                  expect(res.body.recipeData).to.be.an('array');
-                  done();
-                });
-            });
             it(
-              'Should return an error if an unauthenticated user attempts fo fetch recipes',
+              'Should fetch all the user\'s recipe in the application',
+              (done) => {
+                const testUser = { ...users[0] };
+                request.get(`/api/v1/users/recipes?token=${testUser.tokens}`)
+                  .set('Connection', 'keep alive')
+                  .set('Content-Type', 'application/json')
+                  .type('form')
+                  .end((err, res) => {
+                    expect(res.status).to.equal(200);
+                    expect(res.body.message)
+                      .to.equal(`You currently have ${res
+                        .body.recipeData.length} recipe(s)`);
+                    expect(res.body.recipeData).to.be.an('array');
+                    done();
+                  });
+              }
+            );
+            it(
+              `Should return an error if an
+ unauthenticated user attempts fo fetch recipes`,
               (done) => {
                 request.get('/api/v1/users/recipes')
                   .set('Connection', 'keep alive')
@@ -385,8 +386,8 @@ describe('Recipes Endpoint Tests', () => {
                   });
               });
               it(
-                'Should return an error if no' +
-                'parameter is passed in the search',
+                `Should return an error if no
+ parameter is passed in the search`,
                 (done) => {
                   request.get(`${recipesApi}/search?name=`)
                     .set('Connection', 'keep alive')
